try handle closed shadow roots
this doesn't work because the __uvh prop is removed on both ff and chrome. on ff, you can send along the shadow root from the hook before it's sealed in event `details` like it was a few commits ago (?), but this isn't supported on chrome where the details is always null cross-frame. will need to come up with something else like forcing all open instead. keeping this branch/commit to work on maybe another day